Fix: 회원 정보 불러오기 오류 수정
-isLoading으로 렌더링 순서 조정
-mock interface types 폴더로 이동
