+++ src/mocks/handlers.ts
@@ -1,4 +1,5 @@
 import { http, HttpResponse } from 'msw';
+import { IProfileData } from '../types/interface';
 
 export interface ITestData {
   id: number;
@@ -7,15 +8,6 @@ export interface ITestData {
 
 const testData: ITestData[] = [];
 
-/* profile mock interface */
-export interface IProfileData {
-  username: string;
-  phoneNumber: string;
-  profile_image: string;
-  sell_badge: boolean;
-  manner_badge: boolean;
-  star: number;
-}
 /* profile mock data */
 export const profileData: IProfileData = {
   username: '홍길동',
