+++ src/router.tsx
@@ -4,6 +4,7 @@ import Home from './routes/Home';
 import SignIn from './routes/SignIn';
 import SignUp from './routes/SignUp';
 import MyPage from './routes/MyPage';
+import ProfileUpdate from './routes/ProfileUpdate';
 
 export default function Router() {
   return (
@@ -14,6 +15,7 @@ export default function Router() {
         <Route path='/signup' element={<SignUp />} />
       </Route>
       <Route path='/mypage' element={<MyPage />} />
+      <Route path='/mypage/update' element={<ProfileUpdate />} />
     </Routes>
   );
 }
