+++ src/types/interface.ts
@@ -6,3 +6,13 @@ export type FormValueTypes = {
   phoneNumber: number;
   paymentPassword: number;
 };
+
+/* profile mock interface */
+export interface IProfileData {
+  username: string;
+  phoneNumber: string;
+  profile_image: string;
+  sell_badge: boolean;
+  manner_badge: boolean;
+  star: number;
+}
